break down the components
